requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
SQLAlchemy>=1.4
asyncpg>=0.27
alembic>=1.11
authlib>=1.2
python-jose>=3.3.0
pydantic>=2.6
passlib[bcrypt]>=1.7
python-multipart>=0.0.5
psycopg2-binary>=2.9
python-dotenv>=1.0
loguru>=0.7.0